requirements.txt
torch==2.10.0
torchvision==0.25.0
numpy
pandas==2.0.0
matplotlib==3.7.0
scikit-learn==1.2.0
Pillow==9.5.0
tqdm==4.65.0
PyYAML==6.0